refactor(lib): extract shared text metrics
Reading-time and word-count existed three times (markdown.ts, rst.ts,
rst-renderer.ts) with the pacing constants hardcoded in each, so a metrics
fix had to land in three places and the pipelines could silently disagree.
src/lib/text-metrics.ts is now the single owner of the tokenizer, the
Han/Latin token definitions, and the 200 wpm / 300 cpm constants; the
Python-renderer path uses the FromText variants to skip markup stripping
on already-plain text. Heading moves with getHeadings (its producer) and
is type re-exported from markdown.ts until the content/ split.
